libs: vehicle: {ardupilot, mavlink}: Apply necessary changes to create generic mavlink vehicle
